fix: module spec mis-match on Android (new Architecture) (#237)
* fix different type specified from codegen specs

* fix type in impl

+++ android/src/main/java/com/reactnativecommunity/clipboard/ClipboardModule.java
@@ -238,7 +238,7 @@ public class ClipboardModule extends NativeClipboardModuleSpec {
   }
 
   @Override
-  public void removeListeners(double count) {
+  public void removeListeners(int count) {
 
   }
